Updating

+++ src/streams_memkind_memory_task.c
@@ -58,7 +58,7 @@ extern int omp_get_num_threads();
 STREAM_TYPE *a, *b, *c;
 
 
-int stream_memkind_memory_task(benchmark_results *b_results, communicator world_comm, communicator node_comm, int *array_size, int socket, int cache_size, int repeats, char *pmem_path){
+int stream_memkind_memory_task(benchmark_results *b_results, communicator world_comm, communicator node_comm, size_t *array_size, int socket, size_t cache_size, int repeats, char *pmem_path){
 	int			quantum;
 	int			BytesPerWord;
 	int			k;
